second try module9-task1. fixed
